+++ include/morton_tree/reorder.h
@@ -4,7 +4,7 @@
 #ifndef HPCSE2_EX2_REORDER_H
 #define HPCSE2_EX2_REORDER_H
 
-void reorder(const int N, const uint* const keys, const float * const x,const float * const y,const float* const m,const uint* const label,float *xsorted,float *ysorted,float* msorted)
+void reorder(const int N, const uint* const keys, const double * const x,const double * const y,const double* const m,const uint* const label,double *xsorted,double *ysorted,double* msorted)
 {
   #pragma omp parallel for schedule(static)
     for(int i=0;i<N;i++){
